+++ MatrixSDK/MXSession.m
@@ -1770,25 +1770,18 @@ typedef void (^MXOnResumeDone)(void);
     } failure:failure];
 }
 
-- (MXHTTPOperation*)createRoom:(NSString*)name
-                    visibility:(MXRoomDirectoryVisibility)visibility
-                     roomAlias:(NSString*)roomAlias
-                         topic:(NSString*)topic
-                        invite:(NSArray<NSString*>*)inviteArray
-                    invite3PID:(NSArray<MXInvite3PID*>*)invite3PIDArray
-                      isDirect:(BOOL)isDirect
-                        preset:(MXRoomPreset)preset
-                       success:(void (^)(MXRoom *room))success
-                       failure:(void (^)(NSError *error))failure
+- (MXHTTPOperation*)createRoomWithParameters:(MXRoomCreationParameters*)parameters
+                                     success:(void (^)(MXRoom *room))success
+                                     failure:(void (^)(NSError *error))failure
 {
-    return [matrixRestClient createRoom:name visibility:visibility roomAlias:roomAlias topic:topic invite:inviteArray invite3PID:invite3PIDArray isDirect:isDirect preset:preset success:^(MXCreateRoomResponse *response) {
+    return [matrixRestClient createRoomWithParameters:parameters success:^(MXCreateRoomResponse *response) {
 
-        if (isDirect)
+        if (parameters.isDirect)
         {
             // When the flag isDirect is turned on, only one user id is expected in the inviteArray.
             // The room is considered as direct only for the first mentioned user in case of several user ids.
             // Note: It is not possible FTM to mark as direct a room with an invited third party.
-            NSString *directUserId = (inviteArray.count ? inviteArray.firstObject : nil);
+            NSString *directUserId = (parameters.inviteArray.count ? parameters.inviteArray.firstObject : nil);
             [self onCreatedDirectChat:response withUserId:directUserId success:success];
         }
         else