+++ CHANGES.rst
@@ -6,6 +6,8 @@ Improvements:
  * Aggregations: Implement m.reference aggregations, aka thread ([MSC1849](https://github.com/matrix-org/matrix-doc/blob/matthew/msc1849/proposals/1849-aggregations.md)).
  * Cross-Signing: Add a new module, MXCrossSigning, to handle device cross-signing (vector-im/riot-ios/issues/2890).
  * MXRoom: Add a method to retrieve trusted members count in an encrypted room.
+ * MXSession: Add createRoomWithParameters with a MXRoomCreationParameters model class.
+ * MXCrypto: Expose devicesForUser.
 
 Bug fix:
  * MXEventType: Fix Swift refinement.
@@ -14,6 +16,8 @@ Bug fix:
 API break:
  * MXCrypto: the `downloadKeys` method now returns users cross-signing keys.
  * MXDeviceInfo: the `verified` property has been replaced by `trustLevel`.
+ * MXSession & MXRestClient: the `createRoom` method with a long list of parameters
+   has been replaced by `createRoomWithParameters`.
 
 Changes in Matrix iOS SDK in 0.15.2 (2019-12-05)
 ===============================================